合并Controller,并完善部分图书接口

+++ src/main/java/com/example/controller/UploadController.java
@@ -3,7 +3,7 @@ package com.example.controller;
 import com.example.entity.Result;
 import com.example.entity.dto.Account;
 import com.example.entity.vo.request.UpdateAvatarVO;
-import com.example.pojo.Book;
+import com.example.entity.vo.request.Book;
 import com.example.service.AccountService;
 import com.example.service.BookUploadService;
 import com.example.utils.AliOSSUtils;
@@ -68,7 +68,9 @@ public class UploadController {
                 updateAvatarVO.getNewAvatarUrl(), updateAvatarVO.getNewAvatarUuid());
         if (isSuccess) {
             //调用阿里云工具类删除原来的头像
-            aliOSSUtils.DeleteFile(fileAddress);
+            if(fileAddress!=null){
+                aliOSSUtils.DeleteFile(fileAddress);
+            }
             return Result.success(UrlAndUUID);
         } else {
             return Result.error("更新头像失败");
@@ -149,7 +151,6 @@ public class UploadController {
         String bookFileUUID = aliOSSUtils.uploadFile(bookFile,true);
 
         log.info("图书文件更新的云存储地址为：{}", bookFileUUID);
-
         if (bookUploadService.updateBookFile(bookID, bookFileUUID)) {
             //调用阿里云工具类删除原来的文件
             if (originalBookFileAddress != null) {
